requirements.txt
fastapi
uvicorn
python-multipart
tflite-runtime
numpy
pillow
